% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_meta.R
\name{nc_meta}
\alias{nc_meta}
\title{NetCDF minimal metadata}
\usage{
nc_meta(x)
}
\arguments{
\item{x}{NetCDF data source}
}
\description{
Find the dimensions and variables, their lengths and shapes.
}
\examples{
x <- file.path("/rdsi/PRIVATE/raad/data",
"eclipse.ncdc.noaa.gov/pub/OI-daily-v2/NetCDF/2017/AVHRR",
"avhrr-only-v2.20170607_preliminary.nc")
nc_meta(x)
}
